Reorganize some lost functions
