+++ rustworkx/connectivity.pyi
@@ -48,3 +48,13 @@ def cycle_basis(graph: PyGraph, /, root: Optional[int] = ...) -> List[List[int]]
 def articulation_points(graph: PyGraph, /) -> Set[int]: ...
 def biconnected_components(graph: PyGraph, /) -> BiconnectedComponents: ...
 def chain_decomposition(graph: PyGraph, /, source: Optional[int] = ...) -> Chains: ...
+def digraph_find_cycle(
+    graph: PyDiGraph[_S, _T],
+    /,
+    source: Optional[int] = ...,
+) -> EdgeList: ...
+def digraph_complement(graph: PyDiGraph[_S, _T], /) -> PyDiGraph[_S, Optional[_T]]: ...
+def graph_complement(
+    graph: PyGraph[_S, _T],
+    /,
+) -> PyGraph[_S, Optional[_T]]: ...
