+++ rustworkx/rustworkx.pyi
@@ -19,6 +19,7 @@ from typing import Optional, Set, List, Dict, TypeVar, Tuple, Callable, Union
 
 from .centrality import *
 from .connectivity import *
+from .dag_algo import *
 from .isomorphism import *
 from .layout import *
 from .link_analysis import *
@@ -40,31 +41,6 @@ class NegativeCycle(Exception): ...
 class JSONSerializationError(Exception): ...
 class FailedToConverge(Exception): ...
 
-def collect_runs(
-    graph: PyDiGraph[_S, _T],
-    filter_fn: Callable[[_S], bool],
-) -> List[List[_S]]: ...
-def collect_bicolor_runs(
-    graph: PyDiGraph[_S, _T],
-    filter_fn: Callable[[_S], bool],
-    color_fn: Callable[[_T], int],
-) -> List[List[_S]]: ...
-def dag_longest_path(
-    graph: PyDiGraph[_S, _T], /, weight_fn: Optional[Callable[[int, int, _T], int]] = ...
-) -> NodeIndices: ...
-def dag_longest_path_length(
-    graph: PyDiGraph[_S, _T], /, weight_fn: Optional[Callable[[int, int, _T], int]] = ...
-) -> int: ...
-def dag_weighted_longest_path(
-    graph: PyDiGraph[_S, _T],
-    weight_fn: Callable[[int, int, _T], float],
-    /,
-) -> NodeIndices: ...
-def dag_weighted_longest_path_length(
-    graph: PyDiGraph[_S, _T],
-    weight_fn: Callable[[int, int, _T], float],
-    /,
-) -> float: ...
 def digraph_core_number(
     graph: PyDiGraph,
     /,
@@ -73,11 +49,6 @@ def graph_core_number(
     graph: PyGraph,
     /,
 ) -> int: ...
-def digraph_complement(graph: PyDiGraph[_S, _T], /) -> PyDiGraph[_S, Optional[_T]]: ...
-def graph_complement(
-    graph: PyGraph[_S, _T],
-    /,
-) -> PyGraph[_S, Optional[_T]]: ...
 def digraph_all_simple_paths(
     graph: PyDiGraph,
     origin: int,
@@ -106,21 +77,6 @@ def graph_all_pairs_all_simple_paths(
     min_depth: Optional[int] = ...,
     cutoff: Optional[int] = ...,
 ) -> AllPairsMultiplePathMapping: ...
-def find_negative_cycle(
-    graph: PyDiGraph[_S, _T],
-    edge_cost_fn: Callable[[_T], float],
-    /,
-) -> NodeIndices: ...
-def negative_edge_cycle(
-    graph: PyDiGraph[_S, _T],
-    edge_cost_fn: Callable[[_T], float],
-    /,
-) -> bool: ...
-def digraph_find_cycle(
-    graph: PyDiGraph[_S, _T],
-    /,
-    source: Optional[int] = ...,
-) -> EdgeList: ...
 def digraph_union(
     first: PyDiGraph[_S, _T],
     second: PyDiGraph[_S, _T],
@@ -136,13 +92,6 @@ def graph_union(
     merge_edges: bool = ...,
 ) -> PyGraph[_S, _T]: ...
 def is_planar(graph: PyGraph, /) -> bool: ...
-def is_directed_acyclic_graph(graph: PyDiGraph, /) -> bool: ...
-def topological_sort(graph: PyDiGraph, /) -> NodeIndices: ...
-def lexicographical_topological_sort(
-    dag: PyDiGraph[_S, _T],
-    key: Callable[[_S], str],
-    /,
-) -> List[_S]: ...
 def directed_gnm_random_graph(
     num_nodes: int,
     num_edges: int,
@@ -168,7 +117,6 @@ def undirected_gnp_random_graph(
     seed: Optional[int] = ...,
 ) -> PyGraph: ...
 def read_graphml(path: str, /) -> List[Union[PyGraph, PyDiGraph]]: ...
-def transitive_reduction(graph: PyDiGraph, /) -> Tuple[PyDiGraph, Dict[int, int]]: ...
 def digraph_longest_simple_path(graph: PyDiGraph, /) -> Optional[NodeIndices]: ...
 def graph_longest_simple_path(graph: PyGraph, /) -> Optional[NodeIndices]: ...
 def digraph_transitivity(graph: PyGraph, /) -> float: ...
